requirements.txt
streamlit
google-generativeai>=0.8.0
pydantic
python-dotenv
gtts
tqdm
pydub
